add tianracer_script to tianracer_navigation
