+++ tianracer_navigation/script/multi_goals.py
@@ -1,91 +1,82 @@
-#!/usr/bin/env python
-# Copyright 2017 HyphaROS Workshop.
-# Developer: HaoChih, LIN (hypha.ros@gmail.com)
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+#! /usr/bin/env python3
+#  Author: Liangqian Kong <kongliangqian@huawei.com>
 
-import rospy
-import string
-import math
-import time
-import sys
-
-from std_msgs.msg import String
-from move_base_msgs.msg import MoveBaseActionResult
-from actionlib_msgs.msg import GoalStatusArray
+import rclpy
+from rclpy.node import Node
+from rclpy.action import ActionClient
+from action_msgs.msg import GoalStatus
 from geometry_msgs.msg import PoseStamped
+from nav2_msgs.action import NavigateToPose
 
-class MultiGoals:
-    def __init__(self, goalListX, goalListY, retry, map_frame):
-        self.sub = rospy.Subscriber('move_base/result', MoveBaseActionResult, self.statusCB, queue_size=10)
-        self.pub = rospy.Publisher('move_base_simple/goal', PoseStamped, queue_size=10)   
-        # params & variables
-        self.goalListX = goalListX
-        self.goalListY = goalListY
-        self.retry = retry
+
+class MultiGoals(Node):
+    def __init__(self):
+        super().__init__("multi_goals")
+        self.get_goal_paramter()
+        self.retry = self.retry
         self.goalId = 0
-        self.goalMsg = PoseStamped()
-        self.goalMsg.header.frame_id = map_frame
-        self.goalMsg.pose.orientation.z = 0.0
-        self.goalMsg.pose.orientation.w = 1.0
-        # Publish the first goal
-        time.sleep(1)
-        self.goalMsg.header.stamp = rospy.Time.now()
-        self.goalMsg.pose.position.x = self.goalListX[self.goalId]
-        self.goalMsg.pose.position.y = self.goalListY[self.goalId]
-        self.pub.publish(self.goalMsg) 
-        rospy.loginfo("Initial goal published! Goal ID is: %d", self.goalId) 
-        self.goalId = self.goalId + 1 
+        goalMsg = PoseStamped()
+        goalMsg.header.frame_id = self.map_frame
+        goalMsg.header.stamp = self.get_clock().now().to_msg()
 
-    def statusCB(self, data):
-        if data.status.status == 3: # reached
-            self.goalMsg.header.stamp = rospy.Time.now()                
-            self.goalMsg.pose.position.x = self.goalListX[self.goalId]
-            self.goalMsg.pose.position.y = self.goalListY[self.goalId]
-            self.pub.publish(self.goalMsg)  
-            rospy.loginfo("Initial goal published! Goal ID is: %d", self.goalId)              
-            if self.goalId < (len(self.goalListX)-1):
-                self.goalId = self.goalId + 1
-            else:
-                self.goalId = 0 
-                
+        goalMsg.pose.orientation.w = 1.0
+        goalMsg.pose.position.x = self.goalListX[self.goalId]
+        goalMsg.pose.position.y = self.goalListY[self.goalId]
+
+        self.nav_to_pose_client = ActionClient(self, NavigateToPose, 'navigate_to_pose')
+        while not self.nav_to_pose_client.wait_for_server(timeout_sec=1.0):
+            self._logger.info("'NavigateToPose' action server not available, waiting...")
+        goalMsg_nav = NavigateToPose.Goal()
+        goalMsg_nav.pose = goalMsg
+
+        self._send_goal_future = self.nav_to_pose_client.send_goal_async(goalMsg_nav)
+        self._send_goal_future.add_done_callback(self.goal_response_callback)
+        self._logger.info("Initial goal published! Goal ID is: %d" % self.goalId)
+
+    def goal_response_callback(self, future):
+        goal_handle = future.result()
+        if not goal_handle.accepted:
+            self.get_logger().info('Goal rejected :(')
+            return
+        self.get_logger().info('Goal accepted :)')
+        self._get_result_future = goal_handle.get_result_async()
+        self._get_result_future.add_done_callback(self.get_result_callback)
+    
+    def get_result_callback(self, future):
+        status = future.result().status
+        if status == GoalStatus.STATUS_SUCCEEDED:
+            self._logger.info("goal finish! Goal ID is: %d" % self.goalId)
+            self.goalId = self.goalId + 1
 
 
-if __name__ == "__main__":
-    try:    
-        # ROS Init    
-        rospy.init_node('multi_goals', anonymous=True)
+    def get_goal_paramter(self):
+        goalListX = self.get_parameter_or('~goalListX', '[0.229]')
+        goalListY = self.get_parameter_or('~goalListY', '[1.713]')
+        self.map_frame = self.get_parameter_or('~map_frame', 'map')
+        self.retry = self.get_parameter_or('~retry', '1')
 
-        # Get params
-        goalListX = rospy.get_param('~goalListX', '[2.0, 2.0]')
-        goalListY = rospy.get_param('~goalListY', '[2.0, 4.0]')
-        map_frame = rospy.get_param('~map_frame', 'map' )
-        retry = rospy.get_param('~retry', '1') 
-
-        goalListX = goalListX.replace("[","").replace("]","")
-        goalListY = goalListY.replace("[","").replace("]","")
+        goalListX = goalListX.replace("[", "").replace("]", "")
+        goalListY = goalListY.replace("[", "").replace("]", "")
         goalListX = [float(x) for x in goalListX.split(",")]
         goalListY = [float(y) for y in goalListY.split(",")]
 
-        if len(goalListX) == len(goalListY) & len(goalListY) >=2:          
+        if len(goalListX) == len(goalListY) and len(goalListY) >=1:
             # Constract MultiGoals Obj
-            rospy.loginfo("Multi Goals Executing...")
-            mg = MultiGoals(goalListX, goalListY, retry, map_frame)          
-            rospy.spin()
+            self.get_logger().info("Multi Goals Executing...")
         else:
-            rospy.errinfo("Lengths of goal lists are not the same")
-    except KeyboardInterrupt:
-        print("shutting down")
+            self.get_logger().error("Lengths of goal lists are not the same")
 
+        self.goalListX = goalListX
+        self.goalListY = goalListY
+
+
+def main():
+    rclpy.init()
+    rclpy.spin(MultiGoals())
+
+
+
+if __name__ == "__main__":
+    main()
