--- tianracer_script/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/tianracer_script
-[install]
-install_scripts=$base/lib/tianracer_script
